+++ cmd/neofs-cli/internal/client/client.go
@@ -148,21 +148,6 @@ func GetContainer(ctx context.Context, prm GetContainerPrm) (res GetContainerRes
 	return
 }
 
-// IsACLExtendable checks if ACL of the container referenced by the given identifier
-// can be extended. Client connection MUST BE correctly established in advance.
-func IsACLExtendable(ctx context.Context, c *client.Client, cnr cid.ID) (bool, error) {
-	var prm GetContainerPrm
-	prm.SetClient(c)
-	prm.SetContainer(cnr)
-
-	res, err := GetContainer(ctx, prm)
-	if err != nil {
-		return false, fmt.Errorf("get container from the NeoFS: %w", err)
-	}
-
-	return res.Container().BasicACL().Extendable(), nil
-}
-
 // DeleteContainerPrm groups parameters of DeleteContainerPrm operation.
 type DeleteContainerPrm struct {
 	commonPrm